引入redis

+++ src/main/java/cn/edu/szu/Bajie/dto/add/CollectionAddDto.java
@@ -13,7 +13,7 @@ public class CollectionAddDto implements Serializable {
     @ValidatorDiy(message = "收藏类型不合法！",methodNames = "checkCollectType",predicts = CustomizeValidator.class)
     private Integer type;
 
-    private Integer targetId;
+    private Long targetId;
 
     @ValidatorDiy(message = "收藏标志只能是0或者1",methodNames = "checkIsCollected",predicts = CustomizeValidator.class)
     private Integer isCollected;
